+++ cabal.project
@@ -1,0 +1,6 @@
+packages: .
+
+source-repository-package
+    type: git
+    location: https://github.com/byteverse/bytesmith
+    tag: ac29b91728cea68b5107e69133fd78f0d68f76e0
